rename AuthStore to useAuthStore
